Don't hide message if it's NoneType
Otherwise Cura will crash with an AttributeError since NoneType doesn't have a method called 'hide()'.
This _message is never set during the start-up, and so it theoretically doesn't exist. However it does get set during a bunch of methods of this class and those methods get triggered on signals too. So perhaps during this init, it could call one of those functions that sets it to None and the user would see this crash. This should prevent it from crashing.